tests: move list test to be unit tests
Move to a unit test, no need to build this for every platform we have.
