+++ tests/kernel/common/CMakeLists.txt
@@ -18,11 +18,8 @@ target_sources(app PRIVATE
 	src/atomic.c
 	src/byteorder.c
 	src/clock.c
-	src/dlist.c
 	src/intmath.c
 	src/main.c
-	src/slist.c
-	src/sflist.c
 	src/timeout_order.c
 	src/multilib.c
   src/errno.c
