+++ tests/unit/list/testcase.yaml
@@ -1,0 +1,4 @@
+tests:
+  utilities.lists:
+    tags: utils lists
+    type: unit
